refactor: extract ResumableInner and genericize MaLsCh
Adds the fourth composition tier `ResumableInner` (seed + snapshot +
resume) to `core/inner.rs` and implements it for `CmaEs`. The
MA-LSCh framework moves to a generic `MaLsCh<V, LS>` over a
bound-free `MaLsChGenericState<V, C>` (C = chain-slot payload);
`MaLsChCma`/`MaLsChState` become type aliases with the CMA-specific
`new`/`with_inner_lambda` in a specialized impl block, so all
existing call sites compile unchanged. Trajectories verified
bit-identical before/after on the memetic test seeds.

Closes the `ResumableInner` item in TODO.md (removed in a
follow-up docs commit).
