+++ crates/basin/src/solver/cma_es.rs
@@ -634,3 +634,45 @@ where
         Ok((state, None))
     }
 }
+
+impl<V, M, F> crate::core::inner::ResumableInner<V, F> for CmaEs<V, M, F>
+where
+    F: Scalar,
+    V: VectorLen + Clone + ScaleInPlace<F> + std::ops::IndexMut<usize, Output = F>,
+    M: MatrixIdentity,
+{
+    type State = CmaEsState<V, M, F>;
+
+    /// Fresh chain: a new `CmaEs` seeded from `seed` (the prototype's
+    /// own RNG is never drawn), carrying over the prototype's
+    /// `with_lambda` override, plus an isotropic `CmaEsState` at
+    /// `(x, σ = scale)`. `fx` is ignored: `CmaEsState` has no slot for
+    /// a pre-evaluated mean cost (`Solver::init` evaluates `f(m)`
+    /// itself), so priming would change the eval trajectory.
+    fn seed_chain(&self, x: &V, _fx: F, scale: F, seed: u64) -> (Self, Self::State) {
+        let mut cma = CmaEs::new(seed);
+        if let Some(lambda) = self.lambda_override {
+            cma = cma.with_lambda(lambda);
+        }
+        (cma, CmaEsState::new(x.clone(), scale))
+    }
+
+    /// Reset the local iteration counter so the resumed segment starts
+    /// at iter 0; the evolution state (`m`, `σ`, `C`, paths, previous
+    /// generation) persists — that's the chain.
+    fn prepare_resume(&self, state: &mut Self::State) {
+        state.iter = 0;
+    }
+
+    /// The TolX test at `1e-12 ·` the segment's starting σ (Hansen's
+    /// default, made per-segment-relative so it is resume-safe).
+    fn segment_criteria(
+        &self,
+        state: &Self::State,
+    ) -> Vec<Box<dyn crate::core::termination::TerminationCriterion<Self::State>>> {
+        let tol_x = F::from_f64(1e-12).unwrap() * state.sigma();
+        vec![Box::new(crate::core::termination::CmaEsTolerance::new(
+            tol_x,
+        ))]
+    }
+}
